+++ changelog.md
@@ -1,13 +1,15 @@
 ## 1.2.1
 - add `WoodTypeCollection` (similar to vanilla `ColorCollection`)
+- add `BlockModelGenerators.family` overloads with `TexturedModel`
+- add `ExtendedModelProvider.logTextureMapping`
 
 ## 1.2.0
 - add `ExtendedLootSubProvider` (provides common `HolderGetter`s)
 - add `ExtendedRecipeProvider.family` and helper for stonecutting/smelting conversion recipes between families
 - add `ExtendedLanguageProvider.addHumanized` APIs
 - `ExtendedRecipeProvider` takes a `Map<Block, BlockFamily>` and patches vanilla methods to work with custom families 
-- `BlockModelGenerators.familyWithExistingFullBlock` overloads with texture mappings
-- `ExtendedBlockItemTagsProvider` various helpers for BlockFamilies
+- add `BlockModelGenerators.familyWithExistingFullBlock` overloads with `TextureMapping`
+- `ExtendedBlockItemTagsProvider`: various helpers for BlockFamilies
 - remove `CustomLootTableProvider` (use vanilla `LootTableProvider` directly)
 - remove `VanillaCompat.addFlowerPot` (NeoForge can handle this now)
 - disabled tree API (Vanilla tree generation should now be able to do the same)